feat(rheem-profiler): enhance genetic optimization
* allow for multiple individuals
* allow for configuration via Configuration
* add cross-over

+++ rheem-core/src/main/java/org/qcri/rheem/core/api/Configuration.java
@@ -712,7 +712,7 @@ public class Configuration {
         return this.getOptionalDoubleProperty(key).getAsDouble();
     }
 
-    public double getDoubleProperty(String key, long fallback) {
+    public double getDoubleProperty(String key, double fallback) {
         return this.getOptionalDoubleProperty(key).orElse(fallback);
     }
